@@ -18,13 +18,19 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/http"
 	"net/url"
+	"os"
 	"strings"
 	"time"
 
 	"golang.org/x/oauth2"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/client-go/kubernetes"
+	coreV1 "k8s.io/client-go/kubernetes/typed/core/v1"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
 )
 
 // This file contains authentication related functions and structs.
@@ -33,36 +39,106 @@ import (
 var Scopes = []string{"smartRead", "smartWrite"}
 
 type tokenSource struct {
-	token               oauth2.Token
-	cacheFile, clientID string
+	token                oauth2.Token
+	clientID, secretName string
+	secretsClient        coreV1.SecretInterface
 }
 
-func TokenSource(clientID, cacheFile string) oauth2.TokenSource {
-	return oauth2.ReuseTokenSource(nil, newTokenSource(clientID, cacheFile))
+func TokenSource(clientID, secretName, namespace string) oauth2.TokenSource {
+	return oauth2.ReuseTokenSource(nil, newTokenSource(clientID, secretName, namespace))
 }
 
-func newTokenSource(clientID, cacheFile string) *tokenSource {
-	file, err := ioutil.ReadFile(cacheFile)
-	if err != nil {
-		// no file, corrupted, or other problem: just start with an
-		// empty token.
-		return &tokenSource{clientID: clientID, cacheFile: cacheFile}
+func newTokenSource(clientID, secretName, namespace string) *tokenSource {
+	// get kubernetes client
+	var config *rest.Config
+	if _, err := os.Stat("/var/run/secrets/kubernetes.io"); !os.IsNotExist(err) {
+		// we are running inside a kubernetes cluster, use in-cluster config
+		config, err = rest.InClusterConfig()
+		if err != nil {
+			fmt.Println("Unable to create in-cluster config")
+			panic(err.Error())
+		}
+		// ignore namespace value passed in, use namespace of pod
+		ns, err := os.ReadFile("/var/run/secrets/kubernetes.io/serviceaccount/namespace")
+		if err != nil {
+			panic(err.Error())
+		}
+		namespace = string(ns)
+	} else {
+		// we are not in a cluster, expect kube config file in $HOME dir
+		kubeconfig := os.Getenv("HOME") + "/.kube/config"
+		config, err = clientcmd.BuildConfigFromFlags("", kubeconfig)
+		if err != nil {
+			fmt.Println("Unable to create the out-of-cluster config")
+			panic(err.Error())
+		}
 	}
+
+	clientset, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		fmt.Println("Unable to create kubernetes clientset")
+		panic(err.Error())
+	}
+
+	// get secrets client
+	secCli := clientset.CoreV1().Secrets(namespace)
+
+	// ensure the secret exists!
+	secret, err := secCli.Get(context.TODO(), secretName, metav1.GetOptions{})
+	if err != nil {
+		fmt.Printf("(newTokenSource) Unable to get secret '%s' in '%s' namespace", secretName, namespace)
+		panic(err.Error())
+	}
+
+	// check that we have the access and refresh token. if not, start fresh
+	if secret.Data["access_token"] == nil || secret.Data["refresh_token"] == nil {
+		return &tokenSource{clientID: clientID, secretName: secretName, secretsClient: secCli}
+	}
+
+	secret_data := make(map[string]string)
+	secret_data["access_token"] = string(secret.Data["access_token"])
+	secret_data["token_type"] = string(secret.Data["token_type"])
+	secret_data["refresh_token"] = string(secret.Data["refresh_token"])
+	secret_data["expiry"] = string(secret.Data["expiry"])
+
+	secret_data_json, err := json.Marshal(secret_data)
+	if err != nil {
+		fmt.Println("Unable to marshal data from kubernetes secret")
+		panic(err.Error())
+	}
+
 	var tok oauth2.Token
-	err = json.Unmarshal(file, &tok)
+	err = json.Unmarshal(secret_data_json, &tok)
 	if err != nil {
-		// can't unmarshal?  Return an empty token.
-		return &tokenSource{clientID: clientID, cacheFile: cacheFile}
+		panic(err.Error())
 	}
-	return &tokenSource{clientID: clientID, cacheFile: cacheFile, token: tok}
+
+	return &tokenSource{clientID: clientID, secretName: secretName, secretsClient: secCli, token: tok}
 }
 
 func (ts *tokenSource) save() error {
-	d, err := json.Marshal(ts.token)
+	s, err := ts.secretsClient.Get(context.TODO(), ts.secretName, metav1.GetOptions{})
 	if err != nil {
-		return err
+		fmt.Printf("(tokenSource.save) Unable to get secret '%s'", ts.secretName)
+		panic(err.Error())
 	}
-	err = ioutil.WriteFile(ts.cacheFile, d, 0777)
+
+	// insert new token into stringData
+	s.StringData = map[string]string{}
+	s.StringData["access_token"] = ts.token.AccessToken
+	s.StringData["token_type"] = ts.token.TokenType
+	s.StringData["refresh_token"] = ts.token.RefreshToken
+	// Expiry is a time.Time object. Is there a better way to do this?
+	time_string, err := json.Marshal(ts.token.Expiry)
+	if err != nil {
+		panic(err.Error())
+	}
+	// if we don't do this then we have literal quotes in the string!
+	s.StringData["expiry"] = string(time_string)[1 : len(string(time_string))-1]
+
+	// update secret
+	_, err = ts.secretsClient.Update(context.TODO(), s, metav1.UpdateOptions{})
+
 	return err
 }
 
@@ -108,7 +184,7 @@ func (ts *tokenSource) authorize() (*PinResponse, error) {
 		return nil, fmt.Errorf("invalid server response: %v", resp.Status)
 	}
 
-	body, err := ioutil.ReadAll(resp.Body)
+	body, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return nil, fmt.Errorf("error reading response: %s", err)
 	}
@@ -169,7 +245,7 @@ func (ts *tokenSource) getToken(uv url.Values) error {
 		return fmt.Errorf("invalid server response: %v", resp.Status)
 	}
 
-	body, err := ioutil.ReadAll(resp.Body)
+	body, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return fmt.Errorf("error reading response: %s", err)
 	}
@@ -192,6 +268,7 @@ func (ts *tokenSource) getToken(uv url.Values) error {
 }
 
 func (ts *tokenSource) Token() (*oauth2.Token, error) {
+
 	if !ts.token.Valid() {
 		if len(ts.token.RefreshToken) > 0 {
 			err := ts.refreshToken()
@@ -217,9 +294,13 @@ type Client struct {
 // (Application Key).  Use the Ecobee Developer Portal to create the
 // Application Key.
 // (https://www.ecobee.com/consumerportal/index.html#/dev)
-func NewClient(clientID, cacheFile string) *Client {
+func NewClient(clientID, secretName string, namespace ...string) *Client {
+	ns := "default"
+	if len(namespace) > 0 {
+		ns = namespace[0]
+	}
 	return &Client{oauth2.NewClient(
-		context.Background(), TokenSource(clientID, cacheFile))}
+		context.Background(), TokenSource(clientID, secretName, ns))}
 }
 
 // Authorize retrieves an ecobee Pin and Code, allowing calling code to present them to the user
@@ -227,11 +308,15 @@ func NewClient(clientID, cacheFile string) *Client {
 // This is useful when non-interactive authorization is required.
 // For example: an app being deployed and authorized using ansible, which does not support interacting with commands.
 func Authorize(clientID string) (*PinResponse, error) {
-	return newTokenSource(clientID, "").authorize()
+	return newTokenSource(clientID, "", "").authorize()
 }
 
 // SaveToken retreives a new token from ecobee and saves it to the auth cache
 // after a pin/code combination has been added by an ecobee user.
-func SaveToken(clientID string, cacheFile string, code string) error {
-	return newTokenSource(clientID, cacheFile).accessToken(code)
+func SaveToken(clientID, secretName, code string, namespace ...string) error {
+	ns := "default"
+	if len(namespace) > 0 {
+		ns = namespace[0]
+	}
+	return newTokenSource(clientID, secretName, ns).accessToken(code)
 }
