@@ -18,7 +18,7 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/url"
 	"strconv"
 	"strings"
@@ -47,7 +47,7 @@ func (c *Client) UpdateThermostat(utr UpdateThermostatRequest) error {
 		return fmt.Errorf("invalid server response: %v", resp.Status)
 	}
 
-	body, err := ioutil.ReadAll(resp.Body)
+	body, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return fmt.Errorf("error reading body: %v", err)
 	}
@@ -80,7 +80,7 @@ func (c *Client) GetThermostat(thermostatID string) (*Thermostat, error) {
 		IncludeExtendedRuntime: false,
 		IncludeSettings:        false,
 		IncludeSensors:         true,
-		IncludeWeather:         false,
+		IncludeWeather:         true,
 	}
 	thermostats, err := c.GetThermostats(s)
 	if err != nil {
@@ -186,7 +186,7 @@ func (c *Client) get(endpoint string, rawRequest []byte) ([]byte, error) {
 		return nil, fmt.Errorf("invalid server response: %v", resp.Status)
 	}
 
-	body, err := ioutil.ReadAll(resp.Body)
+	body, err := io.ReadAll(resp.Body)
 	if err != nil {
 		return nil, fmt.Errorf("error reading body: %v", err)
 	}
