@@ -21,7 +21,7 @@ import (
 
 	"github.com/golang/glog"
 	homedir "github.com/mitchellh/go-homedir"
-	"github.com/rspier/go-ecobee/ecobee"
+	"github.com/sabowski/go-ecobee-kube/ecobee"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
