@@ -19,7 +19,7 @@ import (
 	"strings"
 
 	"github.com/golang/glog"
-	"github.com/rspier/go-ecobee/ecobee"
+	"github.com/sabowski/go-ecobee-kube/ecobee"
 	"github.com/spf13/cobra"
 )
 
