@@ -16,7 +16,7 @@ package main
 import (
 	goflag "flag"
 
-	"github.com/rspier/go-ecobee/cmd"
+	"github.com/sabowski/go-ecobee-kube/cmd"
 	flag "github.com/spf13/pflag"
 )
 
